+++ pyxel/data_structure/scene.py
@@ -1,4 +1,4 @@
-#  Copyright (c) European Space Agency, 2017, 2018, 2019, 2020, 2021, 2022.
+#  Copyright (c) European Space Agency, 2017.
 #
 #  This file is subject to the terms and conditions defined in file 'LICENCE.txt', which
 #  is part of this Pyxel package. No part of the package, including
@@ -132,8 +132,8 @@ class Scene:
         """
         try:
             from scopesim import Source
-        except ImportError as exc:
-            raise RuntimeError(
+        except ModuleNotFoundError as exc:
+            raise ModuleNotFoundError(
                 "Package 'scopesim' is not installed ! "
                 "Please run command 'pip install scopesim' from the command line."
             ) from exc