+++ pyxel/detectors/detector.py
@@ -1,4 +1,4 @@
-#  Copyright (c) European Space Agency, 2017, 2018, 2019, 2020, 2021, 2022.
+#  Copyright (c) European Space Agency, 2017.
 #
 #  This file is subject to the terms and conditions defined in file 'LICENCE.txt', which
 #  is part of this Pyxel package. No part of the package, including
@@ -53,6 +53,8 @@ class Detector:
         self._signal: Optional[Signal] = None
         self._image: Optional[Image] = None
         self._data: Optional[DataTree] = None
+
+        self._intermediate: Optional[DataTree] = None
 
         # This will be the memory of the detector where trapped charges will be saved
         self._memory: dict = {}
@@ -179,6 +181,14 @@ class Detector:
             raise RuntimeError("'data' not initialized.")
 
         return self._data
+
+    @property
+    def intermediate(self) -> "DataTree":
+        """TBW."""
+        if self._intermediate is None:
+            raise RuntimeError("'intermediate' not initialized.")
+
+        return self._intermediate
 
     def to_xarray(self) -> "xr.Dataset":
         """Create a new ``Dataset`` from all data containers.