+++ pyxel/models/charge_transfer/arctic_cti/models_arctic_vanilla.py
@@ -1,4 +1,4 @@
-#  Copyright (c) European Space Agency, 2017, 2018, 2019, 2020, 2021, 2022.
+#  Copyright (c) European Space Agency, 2017.
 #
 #  This file is subject to the terms and conditions defined in file 'LICENCE.txt', which
 #  is part of this Pyxel package. No part of the package, including
@@ -23,7 +23,7 @@ try:
     import arcticpy as ac
 
     WITH_ARTICPY = True
-except ImportError:
+except ModuleNotFoundError:
     # No 'arcticpy' library
     WITH_ARTICPY = False