+++ pyxel/pipelines/model_group.py
@@ -1,4 +1,4 @@
-#  Copyright (c) European Space Agency, 2017, 2018, 2019, 2020, 2021, 2022.
+#  Copyright (c) European Space Agency, 2017.
 #
 #  This file is subject to the terms and conditions defined in file 'LICENCE.txt', which
 #  is part of this Pyxel package. No part of the package, including
@@ -64,14 +64,14 @@ class ModelGroup:
     def run(
         self,
         detector: "Detector",
-        with_intermediate_steps: bool,
+        debug: bool,
     ):
         """Execute each enabled model in this group.
 
         Parameters
         ----------
         detector : Detector
-        with_intermediate_steps : bool
+        debug : bool
         """
         model: ModelFunction
         for model in self:
@@ -79,70 +79,87 @@ class ModelGroup:
             model(detector)
 
             # TODO: Refactor
-            if with_intermediate_steps:
+            if debug:
                 import xarray as xr
                 from datatree import DataTree
 
                 # Get current absolute time
                 absolute_time = xr.DataArray(
                     [detector.absolute_time],
-                    dims=["time"],
+                    dims="time",
                     attrs={"units": "s"},
                 )
 
                 # Get current Dataset
-                absolute_time = xr.DataArray(absolute_time, attrs={"units": "s"})
+                # absolute_time = xr.DataArray(absolute_time, attrs={"units": "s"})
                 ds: xr.Dataset = detector.to_xarray().assign_coords(time=absolute_time)
 
+                # TODO: Fix this dirty hack
+                if detector._intermediate is None:
+                    detector._intermediate = DataTree()
+
                 # Get datatree parent 'intermediate'
-                intermediate_key: str = "intermediate"
-                if intermediate_key not in detector.data:
-                    datatree_intermediate: DataTree = DataTree(
-                        name="intermediate", parent=detector.data
-                    )
+                datatree_intermediate = detector.intermediate
+
+                if not datatree_intermediate:
+                    datatree_intermediate.name = "intermediate"
                     datatree_intermediate.attrs = {
                         "long_name": (
-                            "Store all intermediate results modified along a pipeline"
+                            "Store all intermediate results "
+                            "modified along a pipeline"
                         )
                     }
                 else:
-                    datatree_intermediate = detector.data[intermediate_key]  # type: ignore
+                    datatree_intermediate = detector.intermediate
 
                 # TODO: Refactor
-                pipeline_key: str = f"idx_{detector.pipeline_count}"
+                pipeline_key: str = f"time_idx_{detector.pipeline_count}"
                 if pipeline_key not in datatree_intermediate:
-                    datatree_one_time: DataTree = DataTree(
-                        name=pipeline_key, parent=datatree_intermediate
+                    datatree_single_time: DataTree = DataTree(
+                        name=pipeline_key,
+                        parent=datatree_intermediate,
                     )
-                    datatree_one_time.attrs = {
+                    datatree_single_time.attrs = {
                         "long_name": "Pipeline for one unique time",
                         "pipeline_count": detector.pipeline_count,
                         "time": f"{detector.absolute_time} s",
                     }
                 else:
-                    datatree_one_time = datatree_intermediate[pipeline_key]  # type: ignore
+                    datatree_single_time = datatree_intermediate[pipeline_key]  # type: ignore
 
                 # TODO: Refactor
                 model_group_key: str = self._name
-                if model_group_key not in datatree_one_time:
+                if model_group_key not in datatree_single_time:
                     datatree_group: DataTree = DataTree(
-                        name=model_group_key, parent=datatree_one_time
+                        name=model_group_key,
+                        parent=datatree_single_time,
                     )
-                    datatree_group.attrs = {"long_name": f"Model group: {self._name!r}"}
+
+                    # TODO: Refactor this ?
+                    # Convert a model group's name to a better string representation
+                    # Example: 'photon_collection' becomes 'Photon Collection'
+                    group_name: str = " ".join(
+                        map(str.capitalize, self._name.split("_"))
+                    )
+                    datatree_group.attrs = {"long_name": f"Model group: {group_name}"}
                 else:
-                    datatree_group = datatree_one_time[model_group_key]  # type: ignore
+                    datatree_group = datatree_single_time[model_group_key]  # type: ignore
 
                 # TODO: Refactor
                 model_key: str = model.name
                 if model_key not in datatree_group:
                     datatree_model: DataTree = DataTree(
-                        name=model_key, parent=datatree_group
+                        name=model_key,
+                        parent=datatree_group,
                     )
-                    datatree_model.attrs = {"long_name": f"Group: {model.name!r}"}
+                    datatree_model.attrs = {
+                        "long_name": f"Model name: {model.name!r}",
+                        "function_name": f"Model function: {model.func.__name__!r}",
+                    }
                 else:
                     datatree_model = datatree_group[model_key]  # type: ignore
 
-                # TODO: Refactor
+                # TODO: Refactor. Is 'last' needed ?
                 last_key: str = "last"
                 if last_key not in datatree_intermediate:
                     last_full_ds: xr.Dataset = xr.zeros_like(ds)
@@ -150,9 +167,12 @@ class ModelGroup:
                     last_full_ds = datatree_intermediate[last_key]  # type: ignore
 
                 for name, data_array in ds.data_vars.items():
-                    previous_data_array = last_full_ds[name]
+                    if name in last_full_ds:
+                        previous_data_array = last_full_ds[name]
 
-                    if not data_array.equals(previous_data_array):
+                        if not data_array.equals(previous_data_array):
+                            datatree_model[name] = data_array
+                    else:
                         datatree_model[name] = data_array
 
                 # datatree_model