+++ pyxel/pipelines/processor.py
@@ -1,4 +1,4 @@
-#  Copyright (c) European Space Agency, 2017, 2018, 2019, 2020, 2021, 2022.
+#  Copyright (c) European Space Agency, 2017.
 #
 #  This file is subject to the terms and conditions defined in file 'LICENCE.txt', which
 #  is part of this Pyxel package. No part of the package, including
@@ -196,12 +196,12 @@ class Processor:
             setattr(obj, att, new_value)
 
     # TODO: Create a method `DetectionPipeline.run`
-    def run_pipeline(self, with_intermediate_steps: bool) -> None:
+    def run_pipeline(self, debug: bool) -> None:
         """Run a pipeline with all its models in the right order.
 
         Parameters
         ----------
-        with_intermediate_steps : bool
+        debug : bool
 
         Notes
         -----
@@ -218,7 +218,7 @@ class Processor:
             self._log.info("Processing group: %r", group_name)
             models_grp.run(
                 detector=self.detector,
-                with_intermediate_steps=with_intermediate_steps,
+                debug=debug,
             )
 
     # TODO: Refactor '.result'. See #524. Deprecate this method ?