+++ tests/unittests/test_fitting.py
@@ -1,6 +1,6 @@
 """Unittests for the 'ModelFitting' class."""
 
-#  Copyright (c) European Space Agency, 2017, 2018, 2019, 2020, 2021, 2022.
+#  Copyright (c) European Space Agency, 2017.
 #
 #  This file is subject to the terms and conditions defined in file 'LICENCE.txt', which
 #  is part of this Pyxel package. No part of the package, including